backend/chatbot/langchain_core/requirements.txt
langchain-community
langchain_text_splitters
langchain_openai
langchain_chroma 
langchain_huggingface
gradio 
python-dotenv 
pypdf
langchain-groq
langchain-core
sentence-transformers
cryptography